Agregados diversos modulos

+++ api/app/Http/Controllers/UserController.php
@@ -8,9 +8,21 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
+    protected $rules = [
+        'name' => 'required|regex:/^[\pL\s\-]+$/u',
+        'username' => 'required|email|max:255|unique:users,email',
+        'password'=> 'required|min:6'
+    ];
+
+    protected $updateRules = [
+        'name' => 'regex:/^[\pL\s\-]+$/u',
+        'username' => 'email|max:255|unique:users,email',
+        'password'=> 'min:6'
+    ];
 
     /**
      * Display a listing of the resource.
@@ -19,7 +31,7 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::all();
+        $users = User::paginate();
 
         return response()->json($users, 200);
     }
@@ -37,7 +49,24 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $user = User::create($request->all());
+        $validator = Validator::make($request->all(), [ 
+            'name' => 'required', 
+            'username' => 'required|email|unique:users', 
+            'password' => 'required',
+        ]);
+
+        if ($validator->fails()) { 
+            return response()->json(['error'=>$validator->errors()], 401);            
+        }
+
+        $input = $request->all();
+        $input['password'] = Hash::make($input['password']);
+        $user = User::create($input);
+
+        //$success['token'] = 
+        $user->createToken('PortalApp')->accessToken;
+        //$success['name'] = $user->name;
+        //return response()->json(['success'=>$success], 200); 
 
         return response()->json($user, 201);
     }
@@ -48,8 +77,10 @@ class UserController extends Controller
      * @param  \App\Center  $center
      * @return \Illuminate\Http\Response
      */
-    public function show(User $user)
+    public function show($id)
     {
+        $user = User::find($id);
+
         return response()->json($user, 200);
     }
 
@@ -60,9 +91,24 @@ class UserController extends Controller
      * @param  \App\Center  $center
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, User $user)
+    public function update(Request $request, $id)
     {
-        $user->update($request->all());
+        $request->validate($this->updateRules);
+
+        $data = $request->all();
+
+        $user = User::find($id);
+
+        $user->update([
+            'name' => $data['name'],
+            'avatar' => $data['avatar'],
+            'username' => $data['username'],
+            'password' => Hash::make($data['password']),
+            'rut' => $data['rut'],
+            'phone' => $data['phone'],
+            'address' => $data['address'],
+            'dob' => $data['dob']
+        ]);
 
         return response()->json($user, 200);
     }
@@ -73,21 +119,25 @@ class UserController extends Controller
      * @param  \App\Center  $center
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User $user)
+    public function destroy($id)
     {
-        $user->delete();
+        $user = User::find($id);
 
-        return response()->json(null, 204);
+        if ($user != null){
+            $user->delete();
+            return response()->noContent();
+        }
+
+        return response()->json(['message' => 'El usuario no existe'], 500);
     }
 
     public function permissions() {
         $permissions = [];
-          foreach (Permission::all() as $permission) {
+        foreach (Permission::all() as $permission) {
             if (Auth::user()->can($permission->name)) {
-              $permissions[] = $permission->name;
+                $permissions[] = $permission->name;
             }
-          }
-          return response()->json($permissions, 200);
-      }
-
+        }
+        return response()->json($permissions, 200);
+    }
 }
